requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-cpf==0.1.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-registration==0.5.6
djangorestframework==3.12.1
drf-writable-nested==0.6.1
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.25.11
